% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_density.R
\name{compute_density}
\alias{compute_density}
\title{Calculate density of correlation.}
\usage{
compute_density(control, correlations, window_number)
}
\arguments{
\item{control}{A list containing random control correlation vectors for a selected window.}

\item{correlations}{A list containing correlations of each window against the top window.}

\item{window_number}{An integer indicating the number of the window for which the density
is to be computed.}
}
\value{
A data frame containing all density data from the control and the window correlations
as named columns.
}
\description{
Computes the density of correlation of the random negative control and the data correlations.
}
\details{
The density data is returned only for the selected window, therefore, \code{compute_density}
needs to be called once per window.

For both the control and the data, \code{compute_density} converts correlations to absolute
values before computing their density.

Density is computed using the \code{density} function, from the \code{stats} package. However,
this function returns a density object, therefore, x and y values of density are subset
using \code{$}.

Since there is more than one control correlation vector for a single window, given the
necessity to randomize the top window several times, the mean of density y values will
be returned instead. Standard error of the mean is also calculated and added and substracted
from the mean to compute the maximum and minimum estimated density.

It should be noted that all control densities are assumed to provide approximately similar
x values, and therefore, only those for the first control density are returned.

All these, together with x and y density values for the data correlations, are bound in
a data frame and returned.

Column names of the returned data frame and their meaning:

\itemize{

\item \code{control_x}: x values of the density object, took from density of one
     the random window's correlations.

\item \code{control_y}: mean y values of density for all random correlations.

\item \code{min_control_y}: \code{control_y} minus standard error of mean.
     \item \code{max_control_y}: \code{control_y} plus standard error of mean

\item \code{data_y}: y values of the density object returned after computing the density
     of the window correlations.

\item \code{data_x}: x values of the density object returned after computing the density
     of the window correlations.
}
}

