% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_histogram.R
\name{compute_histogram}
\alias{compute_histogram}
\title{Compute a quality metric.}
\usage{
compute_histogram(density_data)
}
\arguments{
\item{density_data}{A data frame containing the density data.}
}
\value{
A data frame containing the value of the metric for each window, and values for
error bars to plot them as a histogram.
}
\description{
Computes a statistical metric to assess differences between data distribution
of correlation density and negative control.
}
\details{
The aim of this function is to compute the ratio of correlation density that is non-
coincident with the density distribution of the negative control. Both distributions are
tested to find the section where data density distribution is greater than the control's,
and the differences between both are computed. This metric is divided by the total data
density of correlation.

Graphically, this is the computation of the area where the correlation density distribution
of the data does not overlap with that of the negative control, divided by the total area
under the data correlation density curve. The aim is to determine which windows present
a significant similarity to a random distribution, and therefore are highly affected by
systematic noise, in order to discard them.

The metric is computed by the mean control curve, as well as for the maximum and minimum
curves, obtaining the error values. The returned data frame containis the following three
columns:

\itemize{
     \item \code{hist_value}: the result of computing the metric for the mean control
     curve.

\item \code{error_up}: the value of the metric for the maximum control curve.

\item \code{error_down}: the value of the metric for the minimum control curve.
 }

\strong{Important note:} the \code{compute_histogram} function was designed to operate
on the output of the \code{compute_density} function.
}

