% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Calculate correlations against top window.}
\usage{
correlate(dataset, cor_method)
}
\arguments{
\item{dataset}{A data frame containing all the binned genes.}

\item{cor_method}{A string indicating the type of correlation to use.}
}
\value{
A list containing, on a separate element, the vectors of correlations for every
window against the top.
}
\description{
Calculates correlations of each window against the top, and stores them in a vector.
}
\details{
The input of this function is usually the output of the \code{\link{bin_scdata}} function.

The correlation vector for each window is generated by iterating the genes in the top window
and correlating them to every other window. As a result, there will be as many correlation
values in the vector as genes in the top window. Consequently, top window size greatly impacts
the computational efficiency of the process.

The correlation method argument is passed on to the \code{cor} function, in the \code{stats}
package, and therefore, the same options as this function provides are available. However, it
is adviseable to use pearson correlation, since it presents the most advantageous balance of
result quality and computational efficiency.
}

