% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations_to_densities.R
\name{correlations_to_densities}
\alias{correlations_to_densities}
\title{Transform the correlation table to density distributions of correlation values}
\usage{
correlations_to_densities(df, n = 64, absolute_cc = TRUE)
}
\arguments{
\item{df}{A data frame, usually the output of \code{\link{correlate_windows}}.}

\item{n}{Resolution of the correlation density curve. Default to 64.}

\item{absolute_cc}{Should the function work of absolute value of correlation coefficients?
Default to \code{TRUE} to simplify plots and avoid annoying, non-symmetrical, near 0, shifts of distributions.}
}
\value{
A \code{\link{tibble}} with columns \code{bin}, \code{window}, \code{cor_coef} and \code{density}.
}
\description{
Takes the output of \code{\link{correlate_windows}} and compute density curves of correlation coefficient
for each window comparison.
}
\examples{
#' expMat <- matrix(
    c(1, 1, 5,
      1, 2, 3,
      0, 1, 4,
      0, 0, 2),
    ncol = 3, byrow = TRUE, dimnames = list(paste("gene", 1:4), paste("cell", 1:3))
)

calculate_cvs(expMat) \%>\%
    define_top_genes(window_size = 2) \%>\%
    bin_scdata(window_number = 1) \%>\%
    correlate_windows \%>\%
    correlations_to_densities
}
