% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{extract_top_genes}
\alias{extract_top_genes}
\title{Select top highly expressed genes.}
\usage{
extract_top_genes(dataset, method = c("window_size", "min_expression",
  "mean_expression"), parameter)
}
\arguments{
\item{dataset}{A data frame, containing genes as rows and cells as columns, and where
the mean expression value for each gene has been added as a column.}

\item{method}{A string indicating the method to use when creating the top window. If no
method is indicated, \code{"window_size"} will be used.}

\item{parameter}{An integer. Indicates the numeric parameter to use in the previously
chosen method.}
}
\value{
A list with two elements, both data frames: the generated top window, and
the rest of the genes.
}
\description{
Selects the group of genes in the dataset that will be considered more highly expressed,
the top window, using the filtering method chosen by the user.
}
\details{
There are three selection methods available:

\itemize{

\item \code{"window_size"}: genes are ranked by mean expression, and a subset of the size
     indicated in \code{parameter} is selected from the top.

\item \code{"min_expression"}: genes where all expression values are above a minimum
     expression threshold indicated in \code{parameter} are selected.

\item \code{"mean_expression"}: the \code{mean} column is checked, and all genes with mean
     expression above the threshold indicated in \code{parameter} are selected.
}

There are no restrictions to the \code{parameter} argument, however, the value should
be coherent with the characteristics of the data set provided and the chosen method.

In general, it is adviseable to avoid generating top windows much larger than 250 genes,
to prevent excessively long computation time as well as to preserve the quality of the
analysis, as the top window should only include a subset of reliable values. As a rule,
the bigger the top window is, the more likely is that the reliability of the values is
compromised, given the characteristics of single cell RNA sequencing data.
}

