% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Quality filtering of data.}
\usage{
filter_data(binned_data, histogram, threshold)
}
\arguments{
\item{binned_data}{A data frame, containing the binned genes.}

\item{histogram}{A data frame, containing the quality metric for each window.}

\item{threshold}{A number of type double indicating the quality metric value used to
filter the data.}
}
\value{
A data frame, containing only genes from windows above set minimum quality.
}
\description{
Filters out of the binned data matrix those windows that have a quality metric value
beneath a set threshold.
}
\details{
\code{filter_data} compares the quality metric in every window with the threshold value, and
finds out which windows are above threshold. Then, it subsets the binned data, keeping
only the selected windows.

The output data frame contains gene as rows and cells as columns, but not any other
information concerning the data, since \code{mean}, \code{sd}, \code{bin} and \code{CV}
columns are removed before it is returned.
}

