% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_scatter}
\alias{plot_bins}
\alias{plot_cor_dens}
\alias{plot_histogram}
\alias{plot_scatter}
\title{Plotting the results.}
\usage{
plot_scatter(data, save = FALSE, display = TRUE, path = ".",
  filename = "scatter_plot", format = "png", density = FALSE,
  shape = 16, size = 0.5, alpha = 0.5, ...)

plot_bins(data, save = FALSE, display = TRUE, path = ".",
  filename = "bins_plot", format = "png", shape = 16, alpha = 0.5,
  size = 0.5, ...)

plot_cor_dens(data, save = FALSE, display = TRUE, path = ".",
  filename = NULL, format = "png", window, ribbon_col = "grey",
  ribbon_alpha = 0.2, ..., curve_cols = c("black", "red"))

plot_histogram(data, save = FALSE, display = TRUE, path = ".",
  filename = "histogram", format = "png", title, ylim = 1, ...)
}
\arguments{
\item{data}{A data frame containing the suitably formatted data for the specified plot.}

\item{save}{A logical. When \code{TRUE}, plots will be saved to the indicated path.}

\item{display}{A logical. When \code{TRUE}, plots will be displayed.}

\item{path}{A string indicating a path to save the plot to.}

\item{filename}{A string indicating the name of the plot file.}

\item{format}{A string indicating the file format of the plot.}

\item{density}{Logical. Specifies whether the scatter plot will incorporate scatter
density lines.}

\item{shape, }{size, alpha Numbers specifying aesthetics of the scatter plots, passed on
to the \code{geom_point} function in the \code{ggplot2} package. Shape is an integer, 
size and alpha are doubles.}

\item{...}{Other aesthetic specifications to be passed on to the \code{geom_point}, 
\code{geom_line} or \code{geom_bar} functions.}

\item{window}{An integer. Indicates the window for which the correlation density plot
is being generated by \code{plot_cor_dens}.}

\item{ribbon_col, }{ribbon_alpha Aesthetics for the error space in the correlation density
plot. Color indicated by a string, and alpha indicated by a double.}

\item{curve_cols}{A string vector. Specifies colours of the negative control and data 
curves in the correlation density plots, in that order.}

\item{title}{A string. Contains the title to be set on the histogram plot, normally an
informative title. Recommendation is to specify top window method of selection and size, 
number and size of bins, number of cells in the original dataset.}

\item{ylim}{A double. Sets the limit of the y axis.}
}
\value{
A list, containing the information of the generated plot.
}
\description{
This set of functions generate the summary plots that result from the correlation-based
analysis of your single-cell sequencing data.
}
\details{
Each plotting function is designed to plot the output data of another function in this
package. Associations are as follows:

\itemize{

\item \code{plot_scatter} plots the output of the \code{\link{calculate_cvs}} 
     function, more specifically, only the \code{CV} and \code{mean} columns. However, 
     the plotting function selects these automatically, and the entire data frame 
     should be provided by the user when calling it. 
     
     If running \code{\link{all_calls}}, this can be done by subsetting the results 
     using \code{$cv}.
     
     \item \code{plot_bins} plots the output of the \code{\link{bin_scdata}} function.
     Similarly, the entire data frame should be provided, as the function will select 
     the \code{CV}, \code{mean} and \code{bin} columns automatically from it.
     
     Subset the appropriate results from \code{all_calls} using \code{$binned}.
     
      \item \code{plot_cor_dens} plots the output of the \code{\link{compute_density}} 
      function.It should be noted that this function outputs the density computations 
      for only one window, and therefore, the plotting function must be called again 
      after each computation. 
      
      Subset the appropriate results from \code{all_calls} using 
      \code{$densities[[window number]]}.
      
      \item \code{plot_histogram} plots the output of the \code{\link{compute_histogram}} 
      function. Subset \code{all_calls} results using \code{$histogram}.
}

For the \code{...} argument:

\itemize{

\item \code{plot_scatter} and \code{plot_bins} pass on the arguments to the 
     \code{geom_point} function in \code{ggplot2}.
     
      \item \code{plot_cor_dens} passes arguments on to the \code{geom_line} function in
      \code{ggplot2}.
      
      \item \code{plot_histogram} passes arguments on to \code{geom_bar} function in 
      \code{ggplot2}.
}

For suitable aesthetics, please see documentation of these functions.

\strong{Important note:} calling the plotting functions doesn't enable visualization of
the plot. The plot list that is generated should be saved and then visualized by calling the
\code{plot} function on it.
}

