% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_window_autocor.R
\name{plot_top_window_autocor}
\alias{plot_top_window_autocor}
\title{Utility plot to choose a top_window size}
\usage{
plot_top_window_autocor(sc_data, from = 10, to = 400, by = 2, ...)
}
\arguments{
\item{sc_data}{A tibble, usually the output of \code{link{calculate_cvs}}.}

\item{from}{Minimum size of the top window.}

\item{to}{Maximum size of the top window.}

\item{by}{Size of the steps to walk form \code{from} to \code{to}. See \code{seq}.}

\item{...}{Arguments to be passed to \code{cor}, for example \code{method = "spearman"}}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot mean autocorrelation value of the features of the top window depending
on increasing top window size.
}
\examples{
plot_top_window_autocor(calculate_cvs(scData_hESC))

}
