% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_top.R
\name{randomize_top}
\alias{randomize_top}
\title{Randomize the top window.}
\usage{
randomize_top(dataset)
}
\arguments{
\item{dataset}{A list containing the top window data frame as the first element.}
}
\value{
A matrix containing a sample of randomized values of the top window.
}
\description{
Randomly sorts the values of each row of the supplied object.
}
\details{
\code{randomize_top} generates the result of a single randomization. To create a consistent
negative control, this function should be called a sufficient number of times, using the
average of all randomizations as such.

Randomization takes place one row at a time, sorting the column values and returning the
randomized row to a newly created matrix. Since they are no longer relevant, due to the
sorting, column names are removed.

\strong{Important note:} This function uses the output of \code{\link{extract_top_genes}}
as an input, or any list containing the top subset of genes as the first element. If that
is the case, it should be noted that it will be subset as \code{$topgenes}.
}

