% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scData_hESC.R
\docType{data}
\name{scData_hESC}
\alias{scData_hESC}
\title{Expression data from 32 human embryonic stem cells}
\format{A \code{tibble} with 60,468 rows (genes) and 33 columns (cells):
\describe{
  \item{tracking_id}{The Gencode human encode gene id}
  \item{next 33 columns}{Single embryonic stem cells}
}}
\source{
The example dataset was processed by Mantsoki et al. (2016)
\url{https://www.ncbi.nlm.nih.gov/pubmed/26951854},
data from human Embryonic Stem cells was generated by Yan et al. (2013)\url{https://www.ncbi.nlm.nih.gov/pubmed/23934149}.
}
\usage{
scData_hESC
}
\description{
Expression of 60,468 Gencode gene (in FPKM) from 32 single cell RNAseq
of human embryonic stem cells
}
\keyword{datasets}
