% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_calls.R
\name{all_calls}
\alias{all_calls}
\title{Run the method.}
\usage{
all_calls(data, import = TRUE, max_zeros = 0.5, randomizations = 25,
  top_method = "window_size", filter_parameter = 100,
  bin_method = "window_number", bin_parameter = 30,
  cor_method = "pearson")
}
\arguments{
\item{data}{A file name, path or data frame containing the expression value table.}

\item{max_zeros}{A number of type double. Indicates the maximum proportion
of zeros allowed per row in the raw data.}

\item{randomizations}{An integer. Number of times the top window will be 
randomized to generate a random negative control.}

\item{top_method}{A string. Indicates the method for top window selection.}

\item{filter_parameter}{An integer. Specifies a parameter for the top window 
selection method.}

\item{bin_method}{A string. Indicates the method for binning of the genes.}

\item{bin_parameter}{An integer. Specifies the parameter for the selected 
binning method.}

\item{cor_method}{A string. Indicates the correlation method to use.}

\item{imort}{A logical. Indicates whether raw data should be imported \code{data}
from a file or path (\code{TRUE}) or is in a frame already loaded into R 
(\code{FALSE}).}
}
\value{
A list containing all the results generated during the analysis. Each
element on the list contains the results of a function call, and is named as follows:

\itemize{

     \item \code{raw} contains the raw data table as it was imported.
     
     \item \code{cv} contains the data frame output by \code{calculate_cvs}, which
     includes all the data plus the mean, cv and standard deviation columns.
     
     \item \code{binned} contains all the data as output by the combined call of
     \code{extract_top_genes} and \code{bin_scdata}, which includes the bin number
     column in the data frame.
     
     \item \code{window_correlations} contains the output of \code{correlate}, which
     outputs a list containing the vector of correlation coefficients of each window 
     to the top window as a separate element.
     
     \item \code{controls} contains a nested list with the controls -as many as 
     randomizations- computed for each window, as output by calling \code{randomize_top} and 
     \code{compute_control} once per window.
     
     \item \code{densities} contains a list where the outputs of \code{compute_density}
     for each window are stored as a separate element.
     
     \item \code{histogram} contains a data frame with all the histogram values.
     
     \item \code{info} contains a string in which all the necessary information about
     the analysis is included. This should be used as the histogram title.
}

\strong{Note:} use these names to retrieve information by subsetting using \code{$}.
}
\description{
The \code{all_calls} function calls all the functions that form the package
pipeline, according to the provided parameters.
}
\details{
Arguments for this function are passed on to the each one of the functions in 
the package as indicated in this page, as they are called inside the function 
body.

\code{randomizations} is passed as an argument of the \code{\link{randomize_top}} 
function. Even though more random windows than specified by default (25) can be 
generated, it should be noted that bigger numbers may not make random negative
control any more informative.

\code{max_zeros} is passed as an argument of the \code{\link{calculate_cvs}} 
function. For details on values and meaning, see this object's documentation.
 
When run, \code{all_calls} also displays useful messages that help the user
track analysis progression, as well as information on the operations performed 
(i.e. number or size of bins generated, lowest mean expression value included
in top window, etc.). If the retrieved values do not match what is expected, the
pipeline can be stopped at any point and restarted from the beginning.

By default, the analysis is run creating the \strong{top window} by size, selecting
100 genes, and the \strong{binning} of the rest of the genes is done creating
a specific number of windows (30 by default). For details on these methods and
the parameters, see \code{\link{extract_top_genes}} and \code{\link{bin_scdata}}.

\strong{Important note}: This function is used to run the analysis implemented 
in this package. Unless highly familiar with the method, it is not advised to 
run the different functions individually.
}

