% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{bin_scdata}
\alias{bin_scdata}
\title{Bin genes by mean expression.}
\usage{
bin_scdata(dataset, method = c("window_number", "window_size"), parameter)
}
\arguments{
\item{dataset}{A list, containing the top window generated by \code{extract_top_genes}
as the first element, and the rest of undivided genes as the second.}

\item{method}{A string, indicating the method to be used to bin the genes by mean
expression.}

\item{parameter}{An integer. Indicates the numeric parameter to use in the previously
chosen method. Values are not restricted, but should be coherent with the method of choice.}
}
\value{
A data frame containing the binned genes.
}
\description{
Divides the genes that were not included in the top window in windows of the same size,
by their mean expression.
}
\details{
There are two binning methods available:

\itemize{
     \item \code{"window_number"}: Divides the genes into the number of windows specified in
     \code{parameter}, regardless of their size.

\item \code{"window_size"}: Divides the genes into windows of the size specified in
     \code{parameter}, regardless of the number of windows generated.
}

This function uses the \code{ntile} function, in the \code{dplyr} package to assign a bin
number to each gene based on the value contained in the \code{mean} column, corresponding
to its mean expression. These are then added as a the column \code{bin} using the \code{mutate}
function, also in the \code{dplyr} package.

\strong{Important note:} This function is designed to take the list output by the
\code{extract_top_window} function as an argument, operating only on the second element
of it.

Once the genes in it have been binned, both elements of the list are bound
together in a data frame and returned. The output is similar, but a new column \code{bin}
is added, which indicates the window number assigned to each gene.
}

