% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cvs.R
\name{calculate_cvs}
\alias{calculate_cvs}
\title{Calculate coefficient of variation of genes.}
\usage{
calculate_cvs(data, max_zeros = 0.75)
}
\arguments{
\item{data}{A data frame, containing expression values for each gene as rows, and
expression values for the cells as columns.}

\item{max_zeros}{A double indicating the maximum proportion of zero expression values
allowed per row. The value of this argument must be \eqn{0 => max_zeros > 1}}
}
\value{
A data frame, containing the filtered data and the mean, standard deviation and cv
values for each row.
}
\description{
Calculates the coefficient of variation (CV) of each row in the supplied data table.
}
\details{
Before CV computation, the function 
removes all rows that have a proportion of zeros above the specified
threshold. These will be considered highly disrupted by systematic noise. Removing
them also prevents division by zero when calculating CVs.

The data provided must contain gene names as a column in the first position, as output by the
\code{read_tsv} function in the \code{dplyr} package, and an cell names as column names.
These names will be assigned as the rownames of the output data frame.

In the output, mean, standard deviation and CV are incorporated as new columns in the data
frame, named \code{mean}, \code{sd} and \code{CV}.
}

