% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_histogram.R
\name{compute_histogram}
\alias{compute_histogram}
\title{Compute a quality metric.}
\usage{
compute_histogram(control, correlations, window_number)
}
\arguments{
\item{control}{A data frame containing a subset of random control correlation
vectors, computed for a given window.}

\item{correlations}{A a list containing correlations of each window against 
the top window.}

\item{window_number}{An integer indicating the number of the window for which
the histogram metric is tobe computed.}
}
\value{
A number of type double, namely the difference between the median 
correlation values of the control and the window data.
}
\description{
Computes a statistical metric to assess differences between the distribution
of correlation density in the data and the negative control.
}
\details{
The aim of this function is to compare the median correlation value for both the
mean control curve (see \code{\link{compute_control}}) and the data correlations
(see \code{\link{correlate}}), for a given window, and compute the difference.

This metric reflects the differences between the two distributions, and assesses
the quality of the data in that window. The aim is to determine which windows present
a significant similarity to a random distribution, and therefore are highly affected by
systematic noise, in order to discard them.

\strong{Important note:} the \code{compute_histogram} function was designed to operate
on the output of the \code{\link{compute_control}} and \code{\link{correlate}}
functions.
}

